feat: create file for aircraft state
DT-2068
